Move loadCommand logic to 'which'
